% Generated by build script
\name{osrm_nearest}
\alias{osrm_nearest}
\title{Get nearest road point}
\usage{
osrm_nearest()
}
\arguments{
\item{longitude}{Longitude}
\item{latitude}{Latitude}
\item{profile}{"driving" (default)}
}
\value{tibble: one row with snapped lon/lat, distance, name}
\description{Get nearest road point}
