% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrm.R
\name{osrm_table}
\alias{osrm_table}
\title{Calculate distance/duration matrix}
\usage{
osrm_table(coords, profile = "driving")
}
\arguments{
\item{coords}{Matrix or data frame with lon, lat columns}

\item{profile}{"driving" (default)}
}
\value{
tibble: origin_index, dest_index, duration_s, distance_m
}
\description{
Calculate distance/duration matrix
}
