% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrm.R
\name{osrm_nearest}
\alias{osrm_nearest}
\title{Get nearest road point}
\usage{
osrm_nearest(longitude, latitude, profile = "driving")
}
\arguments{
\item{longitude}{Longitude}

\item{latitude}{Latitude}

\item{profile}{"driving" (default)}
}
\value{
tibble: one row with snapped lon/lat, distance, name
}
\description{
Get nearest road point
}
