% Generated by build script
\name{osrm_table}
\alias{osrm_table}
\title{Calculate distance/duration matrix}
\usage{
osrm_table()
}
\arguments{
\item{coords}{Matrix or data frame with lon, lat columns}
\item{profile}{"driving" (default)}
}
\value{tibble: origin_index, dest_index, duration_s, distance_m}
\description{Calculate distance/duration matrix}
